requirements.txt
python-dotenv==1.0.1
python-keycloak==3.9.1
PyJWT==2.8.0
requests==2.31.0